refactor : 팀 관련 테스트 클래스명 변경 및 패키지 이동
